Add files via upload
A crime has taken place and the detective needs my help. The detective gave me the crime scene report, but I somehow lost it. I vaguely remember that the crime was a murder that occurred sometime on Jan.15, 2018 and that it took place in SQL City. I Started by retrieving the corresponding crime scene report from the police department’s database.

In this project, I used SQL to unravel a murder mystery and helped the detective find the murderer.

I also attached a notepad text and a schema, giving details on how I found the murder.